Refactor TLS 1.3 PSK transport: modularize context handling and remove redundant code
