--- libraries/tacacsrs_networking/src/transport/tls_psk/tls_psk.rs
@@ -1,12 +1,0 @@
-use tokio::net::TcpStream;
-
-use crate::transport::abstractions::Transport;
-
-impl Transport for tokio_openssl::SslStream<TcpStream> {
-    type ReadHalf = tokio::io::ReadHalf<Self>;
-    type WriteHalf = tokio::io::WriteHalf<Self>;
-
-    fn split(self) -> (Self::ReadHalf, Self::WriteHalf) {
-        tokio::io::split(self)
-    }
-}
